Rename area_of_interest.py file
