+++ src/virtualship/expedition/space_time_region.py
